rpc-api: make `system.refresh` return `[(string, error)]` instead of nothing.
The internal interface of refreshing services already returns
`Vec<String>` which stores names of services failed to refresh its
manifest, but we cannot know why it failed. The origin `system.refresh`
RPC method returns `null`, so the return is impossible to be known. By
making this change allow us to know who failed and why it fails by an
RPC invocation.